feat(models): Apply BigInteger PK/FK, SQLAlchemy 2.0 mapped style, add new core models, and resolve circular imports

+++ app/models/events_logs/alert_event.py
@@ -1,5 +1,5 @@
 # app/models/events_logs/alert_event.py
-from sqlalchemy import Column, Integer, DateTime, ForeignKey, Text, Enum
+from sqlalchemy import Column, BigInteger, DateTime, ForeignKey, Text, Enum
 from sqlalchemy.orm import relationship, Mapped, mapped_column # Added Mapped, mapped_column
 from typing import Optional # Added Optional
 from datetime import datetime # Added datetime
@@ -12,18 +12,18 @@ class AlertEvent(Base, TimestampMixin, DeviceFKMixin, AlertRuleFKMixin): # UserF
     이는 시스템의 이상 징후 및 사용자 조치 기록을 추적합니다.
     """
     __tablename__ = "alert_events"
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True) # 알림 이벤트의 고유 ID
-    # device_id는 DeviceFKMixin으로부터 상속받습니다.
-    # alert_rule_id는 AlertRuleFKMixin으로부터 상속받습니다.
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, index=True) # 알림 이벤트의 고유 ID
+    # device_id는 DeviceFKMixin으로부터 상속받습니다. (BigInteger)
+    # alert_rule_id는 AlertRuleFKMixin으로부터 상속받습니다. (BigInteger)
    
     # 명시적으로 외래 키 컬럼 정의
-    user_id: Mapped[Optional[int]] = mapped_column(Integer, ForeignKey('users.id'), nullable=True) # 이벤트를 발생시킨 사용자 ID
+    user_id: Mapped[Optional[int]] = mapped_column(BigInteger, ForeignKey('users.id'), nullable=True) # 이벤트를 발생시킨 사용자 ID
    
     severity: Mapped[str] = mapped_column(Enum('INFO', 'WARNING', 'CRITICAL', name='alert_severity', create_type=False), nullable=False) # 알림의 심각도 ('INFO', 'WARNING', 'CRITICAL')
     message: Mapped[str] = mapped_column(Text, nullable=False) # 알림 메시지 내용
    
     # 명시적으로 외래 키 컬럼 정의
-    acknowledged_by_user_id: Mapped[Optional[int]] = mapped_column(Integer, ForeignKey('users.id'), nullable=True) # 알림을 확인한 사용자 ID
+    acknowledged_by_user_id: Mapped[Optional[int]] = mapped_column(BigInteger, ForeignKey('users.id'), nullable=True) # 알림을 확인한 사용자 ID
    
     acknowledged_at: Mapped[Optional[datetime]] = mapped_column(DateTime(timezone=True), nullable=True) # 알림이 확인된 시간
