+++ app/models/relationships/supported_component_metadata.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Text, UniqueConstraint, Enum
+from sqlalchemy import Column, BigInteger, String, Text, UniqueConstraint, Enum
 from sqlalchemy.orm import relationship, Mapped, mapped_column # Added Mapped, mapped_column
 from typing import Optional
 from app.database import Base
@@ -14,8 +14,8 @@ class SupportedComponentMetadata(Base, TimestampMixin, SupportedComponentFKMixin
         UniqueConstraint('supported_component_id', 'meta_key', name='_component_meta_uc'), # Changed component_id to supported_component_id
     )
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True) # 컴포넌트 메타데이터의 고유 ID
-    # supported_component_id는 SupportedComponentFKMixin으로부터 상속받습니다.
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, index=True) # 컴포넌트 메타데이터의 고유 ID
+    # supported_component_id는 SupportedComponentFKMixin으로부터 상속받습니다. (BigInteger)
     meta_key: Mapped[str] = mapped_column(String(100), nullable=False) # 메타데이터의 키 (예: 'status', 'threshold', 'protocol_version')
     meta_value: Mapped[str] = mapped_column(Text, nullable=False) # 메타데이터의 값 (Text로 저장, 타입에 따라 캐스팅)
     meta_value_type: Mapped[str] = mapped_column(Enum('STRING', 'INTEGER', 'FLOAT', 'BOOLEAN', 'ENUM', 'JSON', name='meta_value_type', create_type=False), nullable=False, default='STRING') # 메타데이터 값의 예상 타입 (e.g., 'BOOLEAN' for online/offline)
